@@ -12,16 +12,16 @@
     "require": {
         "php": ">=5.4.0",
         "doctrine/doctrine-bundle": "^1.6",
-        "doctrine/doctrine-cache-bundle": "^1.2",
         "doctrine/orm": "^2.5",
         "incenteev/composer-parameter-handler": "^2.0",
-        "sensio/distribution-bundle": "^5.0",
+        "sensio/distribution-bundle": "^5.0.19",
         "sensio/framework-extra-bundle": "^3.0.2",
-        "symfony/monolog-bundle": "^3.0.2",
+        "symfony/monolog-bundle": "^3.1.0",
         "symfony/polyfill-apcu": "^1.0",
         "symfony/swiftmailer-bundle": "^2.3.10",
-        "symfony/symfony": "3.2.*",
+        "symfony/symfony": "3.3.*",
         "twig/twig": "^1.0||^2.0",
+        "doctrine/doctrine-cache-bundle": "^1.2",
         "symfony/assetic-bundle": "~2.8",
         "jms/security-extra-bundle": "~1.6",
         "jms/di-extra-bundle": "~1.8",
@@ -33,7 +33,8 @@
         "phpunit/phpunit": "~5.0",
         "squizlabs/php_codesniffer": "3.*",
         "sensio/generator-bundle": "^3.0",
-        "symfony/phpunit-bridge": "^3.0"
+        "symfony/phpunit-bridge": "^3.0",
+        "mockery/mockery": "0.9"
     },
     "scripts": {
         "symfony-scripts": [
@@ -56,7 +57,7 @@
             "@composer test-unit",
             "@composer test-functional"
         ],
-        "cs": "./vendor/bin/phpcs --colors --standard=PSR2 --extensions=php src/SimplyTestable/WebsiteBundle",
+        "cs": "./vendor/bin/phpcs --colors --standard=PSR2 --extensions=php src/SimplyTestable/WebsiteBundle tests/",
         "ci": [
             "@composer cs",
             "@composer test"
