@@ -6,6 +6,7 @@ use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\Yaml\Yaml;
 
 /**
  * This is the class that loads and manages your bundle configuration
@@ -22,7 +23,26 @@ class SimplyTestableWebsiteExtension extends Extension
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
 
-        $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $fileLocator = new FileLocator([
+            __DIR__.'/../Resources/config',
+            __DIR__.'/../Resources/config/plans',
+        ]);
+
+        $loader = new Loader\YamlFileLoader($container, $fileLocator);
         $loader->load('services.yml');
+
+        $planData = Yaml::parse(file_get_contents($fileLocator->locate('plans.yml')));
+        $planDistinctionData = Yaml::parse(file_get_contents($fileLocator->locate('distinctions.yml')));
+
+        $container->setParameter('plans', $planData);
+        $container->setParameter(('plan_distinctions'), $planDistinctionData);
+
+
+        $notFoundRedirectMapData = Yaml::parse(file_get_contents($fileLocator->locate('not_found_redirect_map.yml')));
+
+        $container->setParameter('not_found_redirect_map', $notFoundRedirectMapData);
+
+        $cacheableControllersList = Yaml::parse(file_get_contents($fileLocator->locate('cacheable_controllers.yml')));
+        $container->setParameter('cacheable_controllers', $cacheableControllersList);
     }
 }
