@@ -3,72 +3,43 @@
 namespace Tests\WebsiteBundle\Functional\Command\Sitemap;
 
 use SimplyTestable\WebsiteBundle\Command\Sitemap\BuildCommand;
+use SimplyTestable\WebsiteBundle\Services\KernelParametersService;
+use SimplyTestable\WebsiteBundle\Services\ResourceLocator;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Output\NullOutput;
-use Symfony\Component\Templating\EngineInterface;
 use Tests\WebsiteBundle\Functional\AbstractWebTestCase;
 
 class BuildCommandTest extends AbstractWebTestCase
 {
-    /**
-     * @var string
-     */
-    private $sitemapPath;
-
-    /**
-     * @var string
-     */
-    private $webRootPath;
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function setUp()
-    {
-        parent::setUp();
-        $this->webRootPath = realpath($this->container->get('kernel')->getRootDir() . '/../web');
-        $this->sitemapPath = $this->webRootPath . BuildCommand::SITEMAP_RELATIVE_PATH;
-    }
-
-    public function testGetAsService()
-    {
-        $this->assertInstanceOf(
-            BuildCommand::class,
-            $this->container->get('simplytestable.command.sitemap.build')
-        );
-    }
-
     public function testRun()
     {
-        if (file_exists($this->sitemapPath) && is_file($this->sitemapPath)) {
-            unlink($this->sitemapPath);
+        $sitemapPath = sprintf(
+            '%s%s',
+            $this->container->get(KernelParametersService::class)->getWebDir(),
+            BuildCommand::SITEMAP_RELATIVE_PATH
+        );
+
+        if (file_exists($sitemapPath) && is_file($sitemapPath)) {
+            unlink($sitemapPath);
         }
 
-        $this->assertFileNotExists($this->sitemapPath);
+        $this->assertFileNotExists($sitemapPath);
 
-        /* @var EngineInterface $templating */
-        $templating = $this->container->get('templating');
-
-        $command = new BuildCommand(
-            $this->container->get('router'),
-            $this->container->get('simplytestable.services.resourcelocator'),
-            $this->container->get('kernel')->getRootDir(),
-            $templating
-        );
+        $command = $this->container->get(BuildCommand::class);
 
         $returnCode = $command->run(new ArrayInput([]), new NullOutput());
 
         $this->assertEquals(0, $returnCode);
-        $this->assertFileExists($this->sitemapPath);
+        $this->assertFileExists($sitemapPath);
 
-        $resourceLocator = $this->container->get('simplytestable.services.resourcelocator');
+        $resourceLocator = $this->container->get(ResourceLocator::class);
 
         $urlRoutesSource = file_get_contents(
             $resourceLocator->locate(BuildCommand::SITEMAP_ROUTES_RESOURCE_NAME)
         );
         $urlRoutes = json_decode($urlRoutesSource, true);
 
-        $sitemapContent = file_get_contents($this->sitemapPath);
+        $sitemapContent = file_get_contents($sitemapPath);
         $sitemapDOM = new \DOMDocument();
         $sitemapDOM->loadXML($sitemapContent);
 
