@@ -1,7 +1,8 @@
 <?php
 
-namespace Tests\WebsiteBundle\Functional\Controller;
+namespace Tests\WebsiteBundle\Functional\Services;
 
+use SimplyTestable\WebsiteBundle\Services\ResourceLocator;
 use Tests\WebsiteBundle\Functional\AbstractWebTestCase;
 
 class ResourceLocatorTest extends AbstractWebTestCase
@@ -14,7 +15,7 @@ class ResourceLocatorTest extends AbstractWebTestCase
      */
     public function testLocate($name, $expectedRelativeResourcePath)
     {
-        $resourceLocator = $this->container->get('simplytestable.services.resourcelocator');
+        $resourceLocator = $this->container->get(ResourceLocator::class);
 
         $expectedAbsoluteResourcePath = str_replace(
             '/app',
